requirements.txt
fastapi
uvicorn
openai
python-multipart
pypdf
faiss-cpu
numpy